+++ pkg/wait/conditions.go
@@ -12,6 +12,11 @@ import (
 	cr "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
+type Range struct {
+	Min int
+	Max int
+}
+
 // WaitCondition is a function performing a condition check for if we need to keep waiting
 type WaitCondition func() (done bool, err error)
 
@@ -102,10 +107,28 @@ func DoesResourceExist(ctx context.Context, kubeClient *client.Client, resource 
 	}
 }
 
-// IsNumNodesReady returns a WaitCondition that checks if the number of ready nodes matching the given labels equals or exceeds the expectedNodes value
-func IsNumNodesReady(ctx context.Context, kubeClient *client.Client, expectedNodes int, labels *cr.MatchingLabels) WaitCondition {
+// AreNumNodesReadyWithinRange returns a WaitCondition that checks if the number of ready nodes matching the given labels is within the expected range
+func AreNumNodesReadyWithinRange(ctx context.Context, kubeClient *client.Client, expectedNodes Range, labels *cr.MatchingLabels) WaitCondition {
+	condition := func(readyNodes int) bool {
+		logger.Log("%d nodes ready, expecting between %d and %d", readyNodes, expectedNodes.Min, expectedNodes.Max)
+		return expectedNodes.Min > readyNodes || expectedNodes.Max < readyNodes
+	}
+	return checkNodesReady(ctx, kubeClient, condition, labels)
+}
+
+// AreNumNodesReady returns a WaitCondition that checks if the number of ready nodes matching the given labels equals or exceeds the expectedNodes value
+func AreNumNodesReady(ctx context.Context, kubeClient *client.Client, expectedNodes int, labels *cr.MatchingLabels) WaitCondition {
+	condition := func(readyNodes int) bool {
+		logger.Log("%d nodes ready, expecting %d", readyNodes, expectedNodes)
+		return readyNodes < expectedNodes
+	}
+
+	return checkNodesReady(ctx, kubeClient, condition, labels)
+}
+
+func checkNodesReady(ctx context.Context, kubeClient *client.Client, condition func(int) bool, labels *cr.MatchingLabels) WaitCondition {
 	return func() (bool, error) {
-		logger.Log("Checking for ready control plane nodes")
+		logger.Log("Checking for ready nodes")
 
 		nodes := &corev1.NodeList{}
 		err := kubeClient.List(ctx, nodes, labels)
@@ -123,9 +146,7 @@ func IsNumNodesReady(ctx context.Context, kubeClient *client.Client, expectedNod
 			}
 		}
 
-		logger.Log("%d nodes ready, expecting %d", readyNodes, expectedNodes)
-
-		if readyNodes < expectedNodes {
+		if condition(readyNodes) {
 			return false, nil
 		}